requirements.txt
# Enterprise Content Intelligence Agent - Dependencies

# Core AI/ML
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Observability & Monitoring
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-cloud-trace>=1.6.0
structlog>=23.1.0
prometheus-client>=0.17.0

# Database & Storage
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# MCP Support
mcp>=0.1.0

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Async Support
asyncio>=3.4.3

# Optional: For code execution sandbox
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0